requirements.txt
# LangGraph for agent orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Gemini AI
google-generativeai>=0.8.0

# Database
supabase>=2.9.0
psycopg2-binary>=2.9.9

# Web search
tavily-python>=0.5.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.9.0
